Module sc2::data

Contains useful data exposed through interfaces to the game instance.

Structs

AbilityData
Data about an ability.

BuffData
Buff data.

Effect
Visuals of a persistent ability on the map (eg. PsiStorm).

EffectData
Effect data.

GameSetup
Settings for a game.

ImageData
Data used to sample the current state of the map at certain points.

MapInfo
Info about the terrain.

PlayerResult
Game result tied to a specific player id.

PowerSource
Power source information for Protoss.

Rect
Generic structure to represent a 2D rectangle.

Rect2
2D rectangle represented by two points.

Rect2I
2D integer rectangle represented by two points.

ReplayInfo
Information about a replay file.

ReplayPlayerInfo
Information about a player in a replay.

Score
Score evaluated at the end of a game.

Unit
A unit (could be structure, a worker, or military).

UnitTypeData
Data about a unit type.

UpgradeData
Upgrade data.

Enums

Ability
A list of known StarCraft II abilities

Alliance
Relationship to this player.

Buff
A list of known StarCraft II buffs.

Difficulty
Difficulty setting for built-in StarCraft II AI.

DisplayType
Whether the unit is shown on screen or not.

GameResult
Result of the game.

Map
Different ways of specifying a map.

PlayerSetup
Settings for players.

Race
Race of the player.

UnitType
list of known StarCraft II unit types

Upgrade
A list of known StarCraft II upgrades.

Visibility
Visibility of a point on the terrain.

Type Definitions

Color
Color type for debug commands.

Point2
2D point used to specify location.

Point3
3D point used to specify location.

Point2I
2D integer point used to specify a location.

Tag
Unique tag for a unit instance.

Vector2
2D vector used to specify direction.

Vector3
3D vector used to specify direction.