Struct sc2::data::AbilityData

pub struct AbilityData { /* fields omitted */ }

Data about an ability.

Methods

impl AbilityData

pub fn get_generalized_ability(&self) -> Ability

Get the most generalized id of the ability.

pub fn is_available(&self) -> bool

Indicates whether the ability is available to the current mods/map.

pub fn get_id(&self) -> Ability

Stable ID for the ability.

pub fn get_link_name(&self) -> &str

Catalog (game data xml) name of the ability.

pub fn get_link_index(&self) -> u32

Catalog (game data xml) index of the ability.

pub fn get_button_name(&self) -> &str

Name of the button for the command card.

pub fn get_friendly_name(&self) -> &str

In case the button name is not descriptive.

pub fn get_hotkey(&self) -> &str

UI hotkey.

pub fn get_remap_abilities(&self) -> &[Ability]

Other abilities that can remap to this generic ability.

pub fn get_target(&self) -> Option<AbilityTarget>

Type of target that this ability uses.

pub fn casts_in_minimap(&self) -> bool

Can be cast in the minimap (unimplemented).

pub fn can_autocast(&self) -> bool

Autocast can be set.

pub fn is_building(&self) -> bool

Requires placement to construct a building.

pub fn get_footprint_radius(&self) -> Option<f32>

If the ability is placing a building, give the radius of the footprint.

pub fn is_instant_placement(&self) -> bool

Placement next to an existing structure (an addon like a Tech Lab).

pub fn get_cast_range(&self) -> f32

Range unit can cast ability without needing to approach target.