Module sc2::debug

Contains the public API of the SC2 Debug commands.

Structs

DebugAabb
A debug axis-aligned bounding box defined by two corners.

DebugClient
Debug interface for a game instance.

DebugLine
A debug line defined by a start and end point.

DebugSphere
A debug sphere defined by a point in world space and a radius.

DebugText
Debug text.

Enums

DebugCommand
A debug command for the game.

DebugTextTarget
Target for debugging text.