Enum sc2::AgentSoma

pub enum AgentSoma {
    Init(Init),
    Setup(Setup),
    CreateGame(CreateGame),
    GameCreated(GameCreated),
    JoinGame(JoinGame),
    FetchGameData(FetchGameData),
    StepperSetup(StepperSetup),
    Update(Update),
    SendActions(SendActions),
    SendDebug(SendDebug),
    Step(Step),
    Observe(Observe),
    LeaveGame(LeaveGame),
    Reset(Reset),
}

mediates interactions between the player and the game instance

Variants
Init(Init)

initialize the axon

Setup(Setup)

perform setup queries

CreateGame(CreateGame)

order the game instance to create a game

GameCreated(GameCreated)

game has been created

JoinGame(JoinGame)

join an existing game

FetchGameData(FetchGameData)

order the observer to fetch game data

StepperSetup(StepperSetup)

query the request interval from the player soma

Update(Update)

broadcast game updates to the player soma

SendActions(SendActions)

send any actions for this step to the game instance

SendDebug(SendDebug)

send any debug actions for this step to the game instance

Step(Step)

step the game instance

Observe(Observe)

order the observer to observe the game state

LeaveGame(LeaveGame)

leave the current game

Reset(Reset)

reset the client and re-enter setup state

Methods

impl AgentSoma

pub fn organelle<T>(soma: T) -> Result<Organelle<Sheath<Self>>> where
T: Soma + 'static,
T::Signal: From<Signal>,
T::Synapse: From<Synapse>,
Signal: From<T::Signal>,
Synapse: From<T::Synapse>,

compose an agent organelle to interact with a controller soma