Enum sc2::ErrorKind

pub enum ErrorKind {
    Msg(String),
    Io(Error),
    Ctrlc(Error),
    FutureCanceled(Canceled),
    UrlParse(ParseError),
    Protobuf(ProtobufError),
    Timer(TimerError),
    Tungstenite(Error),
    ExeNotSpecified,
    ExeDoesNotExist(PathBuf),
    AutoDetectFailed(String),
    InvalidMapPath(String),
    MissingRequirement(String),
    InvalidMatch(String),
    ClientOpenFailed(String),
    ClientSendFailed(String),
    ClientRecvFailed(String),
    ClientCloseFailed(String),
    GameErrors(Vec<String>),
    EventAckCanceled(String),
    InvalidProtobuf(String),
    // some variants omitted
}

The kind of an error.

Variants
Msg(String)

A convenient variant for String.

Io(Error)

Link io errors.

Ctrlc(Error)

Link to Ctrl-C errors.

FutureCanceled(Canceled)

Link to futures.

UrlParse(ParseError)

Link to url parse errors.

Protobuf(ProtobufError)

Link to protobuf errors.

Timer(TimerError)

Link to timer errors.

Tungstenite(Error)

Link to tungstenite errors.

ExeNotSpecified

Executable was not supplied to the coordinator.

ExeDoesNotExist(PathBuf)

Executable supplied to the coordinator does not exist.

AutoDetectFailed(String)

Auto-detecting the SC2 installation was unsuccessful.

InvalidMapPath(String)

An invalid map path was supplied to the library.

MissingRequirement(String)

A required field was not provided to a builder.

Often, a builder will have no suitable default for a value. These fields require the user to supply a value. When the builder is finalized, it will check these values and if it is missing a requirement, you should expect this error.

InvalidMatch(String)

Match settings are invalid.

ClientOpenFailed(String)

Client failed to open connection to the game instance.

ClientSendFailed(String)

Client failed to send a message to the game instance.

ClientRecvFailed(String)

Client failed to receive a message from the game instance.

ClientCloseFailed(String)

Client failed to initiate close handshake.

GameErrors(Vec<String>)

Errors received from game instance.

EventAckCanceled(String)

EventAck receiver was dropped or closed.

This should not happen in sc2-rs, but any external libraries with more flexible event subscribers may encounter this problem, and this error will allow them the freedom to deal with it in their own way.

InvalidProtobuf(String)

Invalid protobuf data from game instance.

Methods

impl ErrorKind

pub fn description(&self) -> &str

A string describing the error kind.