Enum sc2::MeleeSoma

pub enum MeleeSoma {
    Init(Init),
    Setup(Setup),
    Launch(Launch),
    PlayerVsPlayer(PlayerVsPlayer),
    PlayerVsComputer(PlayerVsComputer),
    Completed(Completed),
}

soma designed to pit two bots against each other in Sc2 games

Variants
Init(Init)

wait for axon to gather effector, inputs, and outputs

Setup(Setup)

fetch player info in order to decide how many instances it needs

Launch(Launch)

gather instances and game ports, then transition to PVP or PVC

PlayerVsPlayer(PlayerVsPlayer)

coordinate two instances for player vs player

PlayerVsComputer(PlayerVsComputer)

coordinate one instance for player vs the built-in Sc2 AI

Completed(Completed)

melee suite is exhausted and organelle is awaiting shutdown

Methods

impl MeleeSoma

pub fn organelle<L1, L2>(
settings: MeleeSettings<L1, L2>
) -> Result<Organelle<Sheath<Self>>> where
L1: Soma,
L2: Soma,
Signal: From<L1::Signal> + From<L2::Signal>,
Synapse: From<L1::Synapse> + From<L2::Synapse>,
L1::Signal: From<Signal>,
L2::Signal: From<Signal>,
L1::Synapse: From<Synapse>,
L2::Synapse: From<Synapse>,

create the melee organelle