Struct sc2::AbilityData

pub struct AbilityData {
    pub available: bool,
    pub ability: Ability,
    pub link_name: String,
    pub link_index: u32,
    pub button_name: String,
    pub friendly_name: String,
    pub hotkey: String,
    pub remaps_to_ability: Option<Ability>,
    pub remaps_from_ability: Vec<Ability>,
    pub target: Option<AbilityTarget>,
    pub allow_minimap: bool,
    pub allow_autocast: bool,
    pub is_building: bool,
    pub footprint_radius: f32,
    pub is_instant_placement: bool,
    pub cast_range: f32,
}

data about an ability

Fields
available: bool

indicates whether the ability is available to the current mods/map

ability: Ability

stable ID for the ability

link_name: String

catalog (game data xml) name of the ability

link_index: u32

catalog (game data xml) index of the ability

button_name: String

name of the button for the command card

friendly_name: String

in case the button name is not descriptive

hotkey: String

UI hotkey

remaps_to_ability: Option<Ability>

this ability may be represented by this more generic ability

remaps_from_ability: Vec<Ability>

other abilities that can remap to this generic ability

target: Option<AbilityTarget>

type of target that this ability uses

allow_minimap: bool

can be cast in the minimap (unimplemented)

allow_autocast: bool

autocast can be set

is_building: bool

requires placement to construct a building

footprint_radius: f32

if the ability is placing a building, give the radius of the footprint

is_instant_placement: bool

placement next to an existing structure (an addon like a Tech Lab)

cast_range: f32

range unit can cast ability without needing to approach target

Methods

impl AbilityData

pub fn get_generalized_ability(&self) -> Ability

get the most generalized id of the ability