Struct sc2::MeleeBuilder

pub struct MeleeBuilder<P1: MeleeCompetitor + 'static, P2: MeleeCompetitor + 'static> { /* fields omitted */ }

build a Melee coordinator

Methods

impl<P1, P2> MeleeBuilder<P1, P2> where
P1: MeleeCompetitor + 'static,
P2: MeleeCompetitor + 'static,

pub fn new(player1: P1, player2: P2) -> Self

start building a melee soma with the given agent or computer somas

pub fn launcher_settings(self, settings: Launcher) -> Self

the settings for the launcher soma

pub fn one_and_done(self, game: GameSetup) -> Self

play one game with the given settings

pub fn repeat_forever(self, game: GameSetup) -> Self

keep restarting game with the given settings

pub fn update_scheme(self, scheme: UpdateScheme) -> Self

the method of updating the game instance

pub fn break_on_ctrlc(self, flag: bool) -> Self

stop running upon CTRL-C

this is only necessary with Wine. CTRL-C doesn't seem to kill it for some reason by default.

pub fn handle(self, handle: Handle) -> Self

the tokio core handle to use

pub fn create(self) -> Result<Melee> where
P1::Soma: Soma,
P2::Soma: Soma,
<P1::Soma as Soma>::Synapse: Synapse,
<P2::Soma as Soma>::Synapse: Synapse,
<P1::Soma as Soma>::Synapse: From<Synapse> + Into<Synapse>,
<P2::Soma as Soma>::Synapse: From<Synapse> + Into<Synapse>,
<<P1::Soma as Soma>::Synapse as Synapse>::Terminal: From<Terminal> + Into<Terminal>,
<<P2::Soma as Soma>::Synapse as Synapse>::Terminal: From<Terminal> + Into<Terminal>,
<<P1::Soma as Soma>::Synapse as Synapse>::Dendrite: From<Dendrite> + Into<Dendrite>,
<<P2::Soma as Soma>::Synapse as Synapse>::Dendrite: From<Dendrite> + Into<Dendrite>,

build the melee coordinator