Struct sc2::Observation

pub struct Observation { /* fields omitted */ }

state of the game (changes every frame)

Methods

impl Observation

pub fn get_player_id(&self) -> u32

the player id associated with the participant

pub fn get_previous_step(&self) -> u32

the previous game step

pub fn get_current_step(&self) -> u32

the current game step

pub fn get_camera_pos(&self) -> Point2

position of the center of the camera

pub fn get_units(&self) -> &[Rc<Unit>]

a list of all known units at the moment

pub fn get_power_sources(&self) -> &[PowerSource]

all power sources associated with the current player

pub fn get_effects(&self) -> &[Effect]

all active effects in vision of the current player

pub fn get_upgrades(&self) -> &[Upgrade]

all upgrades

pub fn get_minerals(&self) -> u32

current mineral count

pub fn get_vespene(&self) -> u32

current vespene count

pub fn get_food_cap(&self) -> u32

the total supply cap given the players max supply

pub fn get_food_used(&self) -> u32

the total supply used by the player

pub fn get_food_army(&self) -> u32

the total supply consumed by army units alone

pub fn get_food_workers(&self) -> u32

the total supply consumed by workers alone

pub fn get_idle_worker_count(&self) -> u32

the number of workers that currently have no orders

pub fn get_army_count(&self) -> u32

the number of army units

pub fn get_warp_gate_count(&self) -> u32

the number of warp gates owned by the player

pub fn get_larva_count(&self) -> u32

the number of larva owned by the player

pub fn get_creep(&self) -> &ImageData

creep image (sample pixels to find tiles with creep)

pub fn get_visibility(&self) -> &ImageData

visibility image (sample pixels to find visible tiles)

pub fn get_score(&self) -> &Score

detailed current set of scores

pub fn filter_units<F>(&self, filter: F) -> Vec<Rc<Unit>> where
F: Fn(&Unit) -> bool,

filter all units based on a custom condition

pub fn sample_creep(&self, _: Point2) -> bool

check if the given point contains creep

pub fn sample_visibility(&self, _: Point2) -> Visibility

get the visibility of the given point for the current player