Module sc2::data

contains useful data exposed through interfaces to the game instance

Structs

AbilityData
data about an ability

Action
an action (command or ability) applied to a unit or set of units

BuffData
buff data

DebugAabb
a debug axis-aligned bounding box defined by two corners

DebugLine
a debug line defined by a start and end point

DebugSphere
a debug sphere defined by a point in world space and a radius

DebugText
debug text

Effect
visuals of a persistent ability on the map (eg. PsiStorm)

EffectData
effect data

GameSetup
settings for a game

ImageData
data used to sample the current state of the map at certain points

MapInfo
terrain info

PlayerResult
game result tied to a specific player id

PowerSource
power source information for Protoss

Rect
generic structure to represent a 2D rectangle

Rect2
2D rectangle represented by two points

Rect2I
2D integer rectangle represented by two points

ReplayInfo
information about a replay file

ReplayPlayerInfo
information about a player in a replay

Score
score evaluated at the end of a game

Unit
a unit (could be structure, a worker, or military)

UnitTypeData
data about a unit type

UpgradeData
upgrade data

Enums

Ability
list of known StarCraft II abilities

ActionTarget
action target

Alliance
relationship to this player

Buff
list of known StarCraft II buffs

DebugCommand
a debug command for the game

DebugTextTarget
target for debugging text

Difficulty
difficulty setting for built-in StarCraft II AI

DisplayType
whether the unit is shown on screen or not

GameResult
result of the game

Map
different ways of specifying a map

PlayerSetup
settings for players

Race
race of the player

UnitType
list of known StarCraft II unit types

Upgrade
a list of known StarCraft II upgrades

Visibility
visibility of a point on the terrain

Type Definitions

Color
color type for debug commands

Point2
2D point used to specify location

Point3
3D point used to specify location

Point2I
2D integer point used to specify a location

Tag
unique tag for a unit instance

Vector2
2D vector used to specify direction

Vector3
3D vector used to specify direction