Struct sc2::data::UnitTypeData

pub struct UnitTypeData { /* fields omitted */ }

data about a unit type

this data is derived from the catalog (xml) data of the game and upgrades

Methods

impl UnitTypeData

pub fn get_id(&self) -> UnitType

stable unit ID

pub fn get_name(&self) -> &str

unit type name (corresponds to the game's catalog)

pub fn is_available(&self) -> bool

whether this unit is available to the current mods/map

pub fn get_cargo_size(&self) -> u32

number of cargo slots this unit occupies in a transport

pub fn get_mineral_cost(&self) -> u32

cost in minerals to build this unit

pub fn get_vespene_cost(&self) -> u32

cost in vespene to build this unit

pub fn get_attributes(&self) -> &[Attribute]

unit attributes (may change based on upgrades)

pub fn get_movement_speed(&self) -> f32

movement speed of this unit

pub fn get_armor(&self) -> f32

armor of this unit

pub fn get_weapons(&self) -> &[Weapon]

weapons on this unit

pub fn get_food_required(&self) -> f32

how much food this unit requires

pub fn get_food_provided(&self) -> f32

how much food this unit provides

pub fn get_ability(&self) -> Ability

which ability id creates this unit

pub fn get_race(&self) -> Option<Race>

the race this unit belongs to

pub fn get_build_time(&self) -> f32

how long a unit takes to build

pub fn has_minerals(&self) -> bool

whether this unit can have minerals (mineral patches)

pub fn has_vespene(&self) -> bool

whether this unit can have vespene (vespene geysers)

pub fn get_tech_alias(&self) -> &[UnitType]

units this is equivalent to in terms of satisfying tech requirements

pub fn get_unit_alias(&self) -> UnitType

units that are morphed variants of the same unit

pub fn get_tech_requirement(&self) -> UnitType

structure required to build this unit (or any with same tech alias)

pub fn get_require_attached(&self) -> bool

whether tech requirement is an addon